refactor to move param logic to putItem func
